requirements/base.txt
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
httpx==0.23.0
